+++ examples/axum_example/src/main.rs
@@ -5,7 +5,7 @@ use axum::{
     http::StatusCode,
     response::Html,
     routing::{get, get_service, post},
-    AddExtensionLayer, Router, Server,
+    Router, Server,
 };
 use entity::post;
 use entity::sea_orm;
@@ -61,8 +61,8 @@ async fn main() -> anyhow::Result<()> {
         .layer(
             ServiceBuilder::new()
                 .layer(CookieManagerLayer::new())
-                .layer(AddExtensionLayer::new(conn))
-                .layer(AddExtensionLayer::new(templates)),
+                .layer(Extension(conn))
+                .layer(Extension(templates)),
         );
 
     let addr = SocketAddr::from_str(&server_url).unwrap();