+++ sea-orm-cli/src/cli.rs
@@ -1,3 +1,4 @@
+use crate::migration::get_subcommands;
 use clap::{App, AppSettings, Arg, SubCommand};
 
 pub fn build_cli() -> App<'static, 'static> {
@@ -94,8 +95,19 @@ pub fn build_cli() -> App<'static, 'static> {
                 .about("Initialize migration directory")
                 .arg(arg_migration_dir.clone()),
         )
+        .subcommand(
+            SubCommand::with_name("generate")
+                .about("Generate a new, empty migration")
+                .arg(
+                    Arg::with_name("MIGRATION_NAME")
+                        .help("Name of the new migation")
+                        .required(true)
+                        .takes_value(true),
+                )
+                .arg(arg_migration_dir.clone()),
+        )
         .arg(arg_migration_dir.clone());
-    for subcommand in sea_schema::migration::get_subcommands() {
+    for subcommand in get_subcommands() {
         migrate_subcommands =
             migrate_subcommands.subcommand(subcommand.arg(arg_migration_dir.clone()));
     }