+++ sea-orm-cli/template/migration/src/m20220101_000001_create_table.rs
@@ -1,4 +1,4 @@
-use sea_schema::migration::prelude::*;
+use sea_orm_migration::prelude::*;
 
 pub struct Migration;